ResourcesUpdater is IPartCostModifier and only handles resources present in prefab
The first is needed as a part of fixing the total cost calculation of resized
part, as it's total cost is simply the prefab cost + sum of module cost
modifiers. The cost of resources should be implicitly included in the later.
The second is for compatibility with CC and other mods that handle resources
dynamically and thus should include their cost to their own cost modifiers.